requirements.txt
python-dotenv
streamlit

azure-ai-documentintelligence==1.0.2
azure-ai-formrecognizer
